feat(order): order automation
- automate a customer's order

[ Fixes #162188460]
