+++ hirola/front/fixtures/test_order_data.json
@@ -1,0 +1,108 @@
+[
+    {
+        "model": "front.CountryCode",
+        "pk": 1,
+        "fields": {
+          "country_code": 254,
+          "country": "Kenya"
+        }
+      },
+    {
+      "model": "front.User",
+      "pk": 1,
+      "fields": {
+        "email": "timonpumba@gmail.com",
+        "first_name": "timon",
+        "last_name": "pumba",
+        "is_staff": true,
+        "is_active": true,
+        "is_change_allowed": false,
+        "country_code": 1,
+        "phone_number": 722000000
+      }
+    },
+    {
+        "model": "front.PhoneCategory",
+        "pk": 1,
+        "fields": {
+          "phone_category": "iPhone",
+          "category_image": "phone_categories/test_image_6.png"
+        }
+      },
+    {
+    "model": "front.PhoneMemorySize",
+    "pk": 1,
+    "fields": {
+        "abbreviation": "GB",
+        "size_number": 8,
+        "category": 1
+    }
+    },
+    {
+    "model": "front.Currency",
+    "pk": 1,
+    "fields": {
+        "currency_abbreviation": "Kshs.",
+        "currency_long_form": "Kenya Shillings"
+    }
+    },
+    {
+    "model": "front.ItemIcon",
+    "pk": 1,
+    "fields": {
+        "item_icon": "iphone"
+    }
+    },
+    {
+    "model": "front.PhoneList",
+    "pk": 1,
+    "fields": {
+        "category": 1,
+        "phone_name": "iPhone Example1",
+        "currency": 1,
+        "price": 25000.00,
+        "size_sku": 1,
+        "icon": 1,
+        "average_review": 5.0,
+        "main_image": "phones/test_image_5.png"
+    }
+    },
+    {
+        "model": "front.OrderStatus",
+        "pk": 1,
+        "fields": {
+          "status": "Pending"
+        }
+      },
+      {
+        "model": "front.PaymentMethod",
+        "pk": 1,
+        "fields": {
+          "payment_method": 1
+        }
+      },
+
+    {
+        "model": "front.Order",
+        "pk": 1,
+        "fields": {
+          "owner": 1,
+          "phone": 1,
+          "status": 1,
+          "quantity": 2,
+          "total_price": 20000,
+          "payment_method": 1,
+          "date": "2018-12-12"
+        }
+      },
+      {
+        "model": "front.ShippingAddress",
+        "pk": 1,
+        "fields": {
+          "order": 1,
+          "pickup": "Evergreen Center",
+          "location": "Kiambu Road",
+          "recepient": "Nala"
+        }
+      }
+]
